+++ x-pack/test/apm_api_integration/tests/alerts/helpers/wait_for_alerts_for_rule.ts
@@ -6,8 +6,27 @@
  */
 
 import type { Client } from '@elastic/elasticsearch';
+import type {
+  AggregationsAggregate,
+  SearchResponse,
+} from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import pRetry from 'p-retry';
-import { getAlertByRuleId } from './alerting_api_helper';
+import { ApmAlertFields, APM_ALERTS_INDEX } from './alerting_api_helper';
+
+async function getAlertByRuleId({ es, ruleId }: { es: Client; ruleId: string }) {
+  const response = (await es.search({
+    index: APM_ALERTS_INDEX,
+    body: {
+      query: {
+        term: {
+          'kibana.alert.rule.uuid': ruleId,
+        },
+      },
+    },
+  })) as SearchResponse<ApmAlertFields, Record<string, AggregationsAggregate>>;
+
+  return response.hits.hits.map((hit) => hit._source) as ApmAlertFields[];
+}
 
 export async function waitForAlertsForRule({
   es,
